feat: remove boards
