+++ cypress/fixtures/tasks.json
@@ -6,8 +6,7 @@
       "categoryId": 0,
       "date": 1649796466352,
       "updatedAt": 1649796466352,
-      "tags": ["exported"],
-      "board": 1
+      "tags": ["exported"]
     },
     {
       "id": 1649796484700,
@@ -15,9 +14,7 @@
       "categoryId": 1,
       "date": 1649796484700,
       "updatedAt": 1649796484700,
-      "tags": ["feature", "dark"],
-      "board": 1649796474835
+      "tags": ["feature", "dark"]
     }
-  ],
-  "boards": [{ "id": 1649796474835, "title": "Theme" }]
+  ]
 }
